Fix currency conversion
Commit a1b9a67 added the country selection feature. The rates given in
conversionRate.json were in currency/USD but script.js calculates the
USD offset by multiplying the foreign currency offset by the conversion
rate. The result was that converted offsets were incorrect. EUR and GBP
offsets were converted to a too-low USD offset and CAD offsets were
converted to a more generous USD offset. For example, £307.88 was
calculated as $227.83 instead of $418.72.

This commit fixes the problem by reversing the sense of the rates in
conversionRate.json so they are now in USD/currency. Within rounding,
the reciprocals of the old rates were equal to the values I looked up at
around 2025-09-12T18:30+00:00. I chose to change the data over changing
the calculation to avoid any possibility of a division-by-zero.